Refactoring of flask ext tests so that you can modalurize tests
