+++ .gitignore
@@ -7,6 +7,7 @@ venv
 *.rdb
 *.autosave
 *.sqlite
+cache
 
 cache_dir
 benchmarking/node_modules
